fixed data type i var

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int i = 0; /*count nodes*/
+	size_t i = 0; /*count nodes*/
 
 	while (h)
 	{
